+++ src/hooks/index.ts
@@ -1,6 +1,7 @@
 export { usePageSessionSeconds } from "./usePageSessionSeconds";
 export { useLocalHealth } from "./useLocalHealth";
 export type { HealthState } from "./useLocalHealth";
+export { useSupabaseQuotaVersion } from "./useSupabaseQuotaVersion";
 export { useRepositories } from "./useRepositories";
 export { useSessionState } from "./useSessionState";
 export { useTheme } from "./useTheme";